pyproject.toml
[tool.poetry]
name = "blog-api"
version = "0.1.0"
description = ""
authors = ["6Katana9 <katana.st@yandex.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
djangorestframework = "^3.15.1"
djangorestframework-simplejwt = "^5.3.1"
psycopg2-binary = "^2.9.9"
python-decouple = "^3.8"
pillow = "^10.3.0"
drf-yasg = "^1.21.7"
django-filter = "^24.2"
django = "^5.0.6"
setuptools = "^70.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
